+++ CHANGELOG.md
@@ -16,6 +16,8 @@ permalink: /docs/en-US/changelog/
 * Greatly improved database backup and import script checks and output ( #2759 )
 * Faster checks for when packages are installed
 * Use curl for network checks first, falling back to wget
+* New automated Ubuntu EOL checks
+* Optional script to remove provisioner logs older than 1 year
 
 ### Bug Fixes